Add files via upload
The goal of this work is to show that a proper modified very deep model pre-trained on ImageNet for image classification can be used to fit a very small dataset without severe overfitting. In this paper, we explored the possibilities of extracting fire features by trained models to improve the accuracy of the traditional approaches. By evaluating known CNN architectures such as VGG16 and InceptionV3 using transfer learning, a fire detection system was created. To solve the overfitting problem caused by training the network on a limited dataset, we improved the number of available training images using traditional data augmentation techniques. In order to balance the efficiency and accuracy, the model is fine-tuned considering the nature of the target problem and fire data. Experimental results on limited fire datasets reveal the effectiveness of the proposed framework and validate its suitability for detecting fire compared to state-of-the-art methods.
The experimental data sets consist of fire images, and some challenging non-fire images are added into the data sets as a supplement.